reformat packages
