+++ src/test/java/frc/team997/lib/math/IntervalTests.java
@@ -14,11 +14,11 @@ See the GNU General Public License for more details.
 You should have received a copy of the GNU General Public License along with SpartanLib2. 
 If not, see <https://www.gnu.org/licenses/>.
 */
-package frc.team997.lib.Math;
+package frc.team997.lib.math;
 
 import static org.junit.Assert.assertEquals;
 
-import frc.team997.lib.Math.Interval.Inclusion;
+import frc.team997.lib.math.Interval.Inclusion;
 import org.junit.Test;
 
 /** Tests for the Interval class to make sure all logic works properly */
